fix:改变功能位置
